Fix flipper collision
Before, if the ball was moving slowly over a flipper, and you moved it
while the ball was rolling over the first half of it, it would go through
it. This was a consequence of how the flippers rotated off axis using a
hack, rather than implementing it into the system.

Collision resolution now uses getCenter() on rigidbodies when calculating
the radius of the collision point. This function is virtual, and is
overriden by Flipper to return the rotation point.