Add game over and control display screens
When the player reaches game over, a screen will show to display score
and how to restart. When pressing 'p', a similar screen will show to
display the controls.

In addition, the line drawn on spheres to display rotation is not shown
on kinematic spheres, removing the line from bumpers.