Fix sphere rolling rotation
When a sphere would roll up an incline, it would rotate in the direction
gravity is pulling it, causing it to look like it was slipping.

Now, spheres have their angular velocities over-written in a collision
callback set in its constructor that sets the angular velocity to match
the relitive velocity of the contact point.

To do this, the collision callbacks parameters had to be changed to
include a vec2 collision point.